Completed inline editable fields
Everything works with the inline editable fields including client
saving when typeahead is selected or a new vendor/category is typed in
and user presses enter. Pressing enter also works for all text fields
as a save and puts the cursor in the next field. Datepicker also works
as expected